Move TreeNamingService stuff into own package
