+++ plugins/org.yakindu.sct.model.sexec/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Export-Package: org.yakindu.sct.model.sexec,
  org.yakindu.sct.model.sexec.extensions,
  org.yakindu.sct.model.sexec.impl,
  org.yakindu.sct.model.sexec.naming,
+ org.yakindu.sct.model.sexec.naming.tree,
  org.yakindu.sct.model.sexec.transformation,
  org.yakindu.sct.model.sexec.util
 Require-Bundle: com.google.inject,
